Update getting ready fr movement feature
Getting ready to finish adding in the movement feature for the treewidgets and the notebook I added a new function and referenced it within the connections function (of the notebook) and next push will likely be a (not)working version of my movement system. 

And the movement system will be likely be ass because I don't see any better way (or I'm to lazy to find a better way) to do it within Qt. 

Makes me want to rewrite the whole thing in cpp. 

Maybe at some point.