Finishing setup for movement system
The wonky, lack of click and drag (as far as I have found) in the library is bothersome, but I deigned to work around it. 
The connections (and re connections) have been reworked so that they align properly to the movement functions.
Now I just have to write the logic, and reload the new information whenever the use wants to move something!

Oh! And I need a visual representation of what mode the user is currently in, and what they have selected as what they want to move, and where they want to move it, before confirming with a 3rd double click. 

Lots of double clicking and hacky-non-user-friendly UI.